#161 chore: moving factories into the models folder
